+++ arni_gui/src/arni_gui/topic_item.py
@@ -1,17 +1,21 @@
 from abstract_item import AbstractItem
 
+from rospy.rostime import Time
+
 class TopicItem(AbstractItem):
     """A TopicItem represents a specific topic which contains many connections and has attributes like the number of sent messages."""
 
-    def __init__(self, seuid, parent=None):
+    def __init__(self, logger, seuid, parent=None):
         """Initializes the TopicItem.
         
         :param seuid: the seuid of the item
         :type seuid: str
+        :param logger: a logger where to log when special events occur
+        :type logger: ModelLogger
         :param parent: the parent-item
         :type parent: AbstractItem
         """
-        AbstractItem.__init__(self, seuid, parent)
+        AbstractItem.__init__(self, logger, seuid, parent)
         #super(TopicItem, self).__init__(seuid, parent)
         self.__parent = parent
         self._type = "topic"
@@ -36,6 +40,8 @@ class TopicItem(AbstractItem):
 
         for item in self.__rated_attributes:
             self._add_rated_data_list(item)
+
+        self._logger.log("info", Time.now(), seuid, "Created a new TopicItem")
 
 
     def execute_action(self, action):
@@ -85,3 +91,11 @@ class TopicItem(AbstractItem):
 
     def get_short_data(self):
         return "TopicItem"
+
+    def can_execute_actions(self):
+        """
+        This item cannot execute actions, so it returns False
+
+        :return: False
+        """
+        return False