+++ rqt_arni_gui_detail/src/rqt_arni_gui_detail/arni_gui_detail.py
@@ -66,11 +66,11 @@ class ArniGuiDetail(Plugin):
         :param item: the double-clicked item
         :type item: QModelIndex
         """
-        print(type(index))
-        print(index.isValid())
-        print(type(index.internalPointer()))
+        #print(type(index))
+        #print(index.isValid())
+        #print(type(index.internalPointer()))
 
-        self.__selection_widget.set_selected_item(index)
+        self.__selection_widget.set_selected_item(self.__tree_widget.item_tree_view.itemFromIndex(index))
 
 
     def shutdown_plugin(self):